Revert "Add telemetry push for ATOM THREADNETWORK_TELEMETRY_DATA..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/84cdc423-d5b0-4176-9221-fd99b5d1be98). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2dd2fb1c5ab2de766299537749076a7a330efa9e