Revert "Add telemetry push for ATOM THREADNETWORK_TELEMETRY_DATA..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8fdcab87-8c2e-4c75-9a2b-ab6db040d0b6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2695701aad5bcf25179d94346ed5526a5ca559ce