Revert "Add telemetry push for ATOM THREADNETWORK_TELEMETRY_DATA..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9a822998-d93e-47f3-92f4-6beaaf66cdc2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I41d2867a6fda63adcf29f9f23d543453756197e9