Revert "Add telemetry push for ATOM THREADNETWORK_TELEMETRY_DATA..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/da8641bb-ee4c-442d-909b-e5d6ec29e4c3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I23157ac877702e04b01e66387cff43e98cbce4f7