Revert "Add telemetry push for ATOM THREADNETWORK_TELEMETRY_DATA..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/99e608bf-19cc-415c-860e-8f9bd9c5596f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I60305735cac90d71adafc29c6ab7595254ac7081