[Thread] add FakeOtDaemon for testing
The IOtDaemon.aidl API is used by ThreadNetworkControllerService, but
the real IOtDaemon implementation can't be used in unittests. Instead of
mocking the IOtDaemon logics everywhere in the unittest methods, here
creates a fake IOtDaemon implementation which has the same (but
customisable) behavior as the real OtDaemon server. This FakeOtDaemon
can then be used for ThreadNetworkControllerService unittests.

Bug: 317555104
Change-Id: I6e079362fc08f24935edc47f1d3ccf4964a3677e