[Thread] Add IOtDaemon setThreadEnabled() API
The setThreadEnabled() API can be used to enable/disable Thread.
When disabling Thread if the device has joined Thread network, it first
detaches from the network without erasing the Thread dataset.
Concurrent calls to the API are queued to make sure each is completed or
error is reported.

Added onThreadEnabledChanged() to IOtDaemonCallback to notify Thread
enabled state changes.

Bug: 299243765

Test: atest OtDaemonUnitTests

Change-Id: I964641ba5a4dab3d4afe78d84603c1aec8d9ccd1