[Android] run IOtDaemon functions in task runner
The IOtDaemon functions will be called when the function `select()`
detects that the Binder fd is readable. The Radio Spinel calls
`select()` to wait for the Spinel response after it sends the Spinel
command. There is a chance that the IOtDaemon function is called
when the Radio Spinel is waiting for the response, and the IOtDaemon
functions may call the Radio Spinel to send a Spinel command to the
RCP and wait for the response. The previous Spinel response will be
received fist, but the Radio Spinel called by IOtDaemon functions
thinks the received response is the response it expected to receive.
Then the sequence number mismatch issue happens and the Thread stack
is crashed.

This commit makes all IOtDaemon functions run in the task runner to
guarantee the Thread functions won't be called when the Radio Spinel
is waiting for the response.

Bug: b/326017063
Test: Run `stop ot-daemon` for multiple times, the Thread stack
      could restarted and no Thread stack crash is found.

Change-Id: If4f90dd84ff987fdd8342ed61de99b85b8717e61