[Thread] add command to force stop ot-daemon
The Thread HAL supports only one client for a single Thread chip, so
ot-daemon needs to be force stopped in VTS tests because the testapp
needs to connect to the HAL service to perform tests.

This commit adds a new terminate() API to ot-daemon to stop itself.

Bug: 328538612
atest: atest ThreadNetworkIntegrationTests:android.net.thread.ThreadNetworkShellCommandTest
Change-Id: Ic6000eaf2a912e3869048efdc0816b217b81e970