Revert "[meshcop] support runtime custom vendor and product name..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/27f0f102-478a-401a-8e19-c247602799f3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0a48de66643401abd17105b55c2f6bf426e81659