Revert "[Thread] support customizing meshcop TXT entries"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/27f0f102-478a-401a-8e19-c247602799f3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0597042eb6a00a3141aa9c136ff9cb2dbbf93d6a