Revert "[Thread] use the former leave() implementation to work a..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ee3542bb-171d-4807-ba20-cb878aeea61e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1e938b4541688f96d925908e8e39dd667377359f