Revert "[Thread] use the former leave() implementation to work a..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/99f60ebc-1249-4361-b08a-56b7c26ad006). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2a0d06ca840e50a0a3489d4b37124ad7f34f5aff