Revert "[Thread] use the former leave() implementation to work a..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a49f3c6e-269e-4eec-a6c4-4d36febd2908). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3b0d68738aac391c70cfdf0ad38358eaa0d941ce