Revert "[Thread] use the former leave() implementation to work a..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e6ecef56-e3e8-4d67-9dc5-27ef2b3a7dcd). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib5dcd7ca9817e6406009be6ad4f0abe17887bb90