Revert "[Thread] use the former leave() implementation to work a..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7c8ca488-12e6-47b1-b7e2-c36cb9319e3e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8a802245a7e9e04cb7dd35b31531785db8872599