Revert "[Thread] use the former leave() implementation to work a..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2c2fdd54-76d0-4845-9ada-9ece40e4105b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia30c53181f3310e73fd6d16980981abe8d098407