Revert "Revert "[Thread] use the former leave() implementation t..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/dea3ce21-92a5-4900-bd1e-aa0cae307c15). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I05d477b1ea9a0c1c9b465bcdf2c4ba5d28a358f0