Revert "Add missing macro definitions in ot-daemon."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/09e14e6d-e262-41d0-bede-d1fc3fa244b5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I23cacd65dfdbc759a8308263485774cc72101ef4