Revert "Fix onThreadEnableStateChanged issue."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1e9a2978-1c97-4d24-9168-cd6324526414). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I18033bc1bf0d8417a4799e1fdf0fcef5b472788f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com